Overlays remain under the legend and probability distribution
