+++ frontend/src/MapWrapper.tsx
@@ -2,7 +2,7 @@ import { skipToken } from '@reduxjs/toolkit/dist/query'
 import { useMemo, useRef } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import DataDescription from './DataDescription'
-import DataProcessor from './DataProcessor'
+import DataProcessor, { getDomain } from './DataProcessor'
 import DataSourceDescription from './DataSourceDescription'
 import EmptyMap from './EmptyMap'
 import FullMap from './FullMap'
@@ -10,14 +10,50 @@ import { DataQueryParams, useGetDataQuery } from './MapApi'
 import MapControls from './MapControls'
 import MapTitle, { EmptyMapTitle } from './MapTitle'
 import MapTooltip from './MapTooltip'
-import { MapVisualization, MapVisualizationId } from './MapVisualization'
+import { MapType, MapVisualization, MapVisualizationId } from './MapVisualization'
 import css from './MapWrapper.module.css'
 import Overlays from './Overlays'
 import { clickMap, selectMapTransform, selectSelections, stateId } from './appSlice'
 import { RootState } from './store'
+import Legend from './Legend'
+import { getLegendFormatter, getUnitString } from './Formatter'
+import Color from './Color'
+import ProbabilityDensity from './ProbabilityDensity'
 
 export const ZOOM_TRANSITION = { transition: 'all 0.5s cubic-bezier(0.4, 0, 0.2, 1)' }
 
+export const getLegendTitle = (selectedMaps: MapVisualization[], isNormalized: boolean) => {
+    const dataDefinition = selectedMaps[0]
+    const unitString = getUnitString({ units: dataDefinition.units, isNormalized })
+
+    if (isNormalized) {
+        if (selectedMaps.some((value) => value.subcategory === 1)) {
+            return selectedMaps.length > 1 ? 'Combined Relative Risk' : 'Relative Risk'
+        }
+        return 'Scaled Value'
+    }
+    return unitString
+}
+
+function getPdfDomain(selectedMaps: MapVisualization[]) {
+    const firstSelection = selectedMaps[0]
+    if (firstSelection === undefined) {
+        return undefined
+    }
+
+    return firstSelection.pdf_domain
+}
+
+function shouldShowPdf(selectedMaps: MapVisualization[], isNormalized: boolean) {
+    const firstSelection = selectedMaps[0]
+    if (selectedMaps[0] !== undefined && selectedMaps[0].show_pdf === false) {
+        return false
+    }
+    if (isNormalized) {
+        return selectedMaps.length > 1
+    }
+    return firstSelection !== undefined && firstSelection.map_type === MapType.Choropleth
+}
 function MapWrapper({
     allMapVisualizations,
     isNormalized,
@@ -71,6 +107,8 @@ function MapWrapper({
     )
     const dataSource =
         maps[0] && selections[0] ? maps[0].sources[selections[0].dataSource] : undefined
+    const getLegendTicks = (selectedMaps: MapVisualization[], isNormalized: boolean) =>
+        isNormalized ? undefined : selectedMaps[0].legend_ticks
 
     if (map === undefined) {
         return <p>Loading</p>
@@ -116,6 +154,46 @@ function MapWrapper({
                         <EmptyMap map={map} transform={transform} />
                     )}
                     <Overlays />
+                    {processedData &&
+                        (() => {
+                            // JavaScript goes here
+                            const legendTitle = getLegendTitle(maps, isNormalized)
+                            const legendFormatter = getLegendFormatter(maps, isNormalized)
+                            const domain = getDomain(processedData)
+                            const colorScheme = Color(isNormalized, detailedView, maps[0], domain)
+                            const getArrayOfData = () =>
+                                Array.from(processedData.valueSeq()).filter(
+                                    (value) => value !== undefined
+                                ) as number[]
+                            const legendTicks = getLegendTicks(maps, isNormalized)
+                            return (
+                                <>
+                                    <Legend
+                                        title={legendTitle}
+                                        colorScheme={colorScheme}
+                                        tickFormat={legendFormatter}
+                                        ticks={legendTicks}
+                                        showHighLowLabels={isNormalized}
+                                        x={map.region === 'World' ? 0 : 875}
+                                        y={map.region === 'World' ? 502 : 500}
+                                        width={290}
+                                        height={60}
+                                    />
+                                    {shouldShowPdf(maps, isNormalized) && (
+                                        <ProbabilityDensity
+                                            data={getArrayOfData()}
+                                            map={maps[0]}
+                                            xRange={getPdfDomain(maps)}
+                                            formatter={legendFormatter}
+                                            continuous={detailedView}
+                                            shouldNormalize={isNormalized}
+                                            width={290}
+                                            height={200}
+                                        />
+                                    )}
+                                </>
+                            )
+                        })()}
                 </svg>
                 {map && (
                     <MapControls data={processedData} isNormalized={isNormalized} maps={maps} />
